Update Start
Importing `ArrayList` allows the program to utilize this class for managing a list of `Appointment` objects efficiently. Implement auxiliary methods under the main method in the Assign One class to demonstrate the use of classes and objects to manage appointment information and healthcare professionals. Several instances of GeneralPractitioner and OtherHealthProfessional were created in the main method, and their printDetails method was called to print detailed information. Then, several appointments were created using the 'create Appointment' method and added to the 'appointments' list. Then, all existing appointment information was printed using the printExistingAdmissions method. Finally, a reservation was cancelled using the cancelBooking method, and the corresponding reservation was searched and removed based on the patient's phone number.